pyproject.toml
[tool.poetry]
name = "kilke-python-boilerplate"
version = "0.1.0"
description = ""
authors = ["Kalle Ilves <kalleilv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.5"
requests = "^2.24.0"
python-dotenv = "^0.15.0"
socketclusterclient = "^1.3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
